+++ __init__.py
@@ -5,12 +5,11 @@ from homeassistant.core import HomeAssistant
 from homeassistant.const import Platform
 
 from .hub import Hub
+type HubConfigEntry = ConfigEntry[Hub]
 
 # List of platforms to support. There should be a matching .py file for each,
 # eg <cover.py> and <sensor.py>
-PLATFORMS = [Platform.SENSOR, Platform.SWITCH]
-
-type HubConfigEntry = ConfigEntry[Hub]
+PLATFORMS = [Platform.SWITCH]
 
 async def async_setup_entry(hass: HomeAssistant, hub_config_entry: HubConfigEntry) -> bool:
     """Set up from a config entry."""
