+++ api.py
@@ -5,6 +5,7 @@ from typing import Any
 import logging
 from aiohttp import ClientSession
 from aiohttp.client import _RequestOptions
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -17,7 +18,7 @@ class Route:
     """API route configuration (method and URI path)"""
     def __init__(self, method: str, uri: str, value_path: ValuePath = None) -> None:
         self.method: str = method
-        self.uri: str = uri,
+        self.uri: str = uri
         self.value_path: ValuePath = value_path
 
     def parse_response(self, response_value: Any) -> Any:
@@ -36,13 +37,20 @@ class Route:
 
 class ApiClient:
     """Sends HTTP requests."""
-    def __init__(self, host, headers) -> None:
-        self.host = host
-        self.headers = headers
+    def __init__(self, host: str, api_key: str) -> None:
+        self.host: str = host
+        self.api_key: str = api_key
+        self.headers = {"x-api-key": api_key}
+
 
     async def send(self, route: Route, request_options: _RequestOptions = None) -> Any:
         """Sends the configured HTTP request"""
+
         url = self.host + route.uri
+
+        if request_options is None:
+            request_options = {}
+
         session: ClientSession = ClientSession()
 
         _LOGGER.debug(
@@ -68,35 +76,41 @@ class ApiClient:
         finally:
             await session.close()
 
-class Api():
-    """Sends HTTP requests and parses the response."""
+# class Api():
+#     """Sends HTTP requests and parses the response."""
 
-    def __init__(self, host: str, api_key: str) -> None:
-        """Init hub for IMMICH API"""
+#     def __init__(self, host: str, api_key: str) -> None:
+#         """Init hub for IMMICH API"""
 
-        self.host = host
-        self.api_key = api_key
-        self.headers = {"x-api-key": api_key}
+#         self.host: str = host
+#         self.api_key: str = api_key
+#         self.headers = {"x-api-key": api_key}
 
-    async def send(self, route: Route, request_options: _RequestOptions = None):
-        """Sends the configured HTTP request"""
-        url = self.host + route.uri
-        async with ClientSession() as session:
+#     async def send(self, route: Route, request_options: _RequestOptions = None):
+#         """Sends the configured HTTP request"""
+#         url = f"{self.host}{route.uri}"
+#         async with ClientSession() as session:
 
-            _LOGGER.debug(
-                "Request: Method: %s URL: %s kwargs: %s",
-                route.method,
-                url,
-                request_options
-            )
+#             _LOGGER.debug(
+#                 "Request: Method: %s URL: %s kwargs: %s",
+#                 route.method,
+#                 url,
+#                 request_options
+#             )
 
-            async with session.request(
-                method = route.method,
-                url = url,
-                headers = self.headers,
-                **request_options
-            ) as response:
-                _LOGGER.debug("Response: %s", response)
-                response.raise_for_status()
-                response_data = await response.json()
-                return route.parse_response(response_data)
+#             async with session.request(
+#                 method = route.method,
+#                 url = url,
+#                 headers = self.headers,
+#                 **request_options
+#             ) as response:
+#                 _LOGGER.debug("Response: %s", response)
+#                 response.raise_for_status()
+#                 response_data = await response.json()
+#                 return route.parse_response(response_data)
+
+class Listener:
+    """ Used to rute responses to callbacks  """
+    def __init__(self, coordinator: DataUpdateCoordinator, value_path: ValuePath) -> None:
+        self.coordinator: DataUpdateCoordinator = coordinator
+        self.value_path: ValuePath = value_path
