+++ const.py
@@ -1,0 +1,3 @@
+""" Constants """
+
+DOMAIN = "immich_api"
