--- constants.py
@@ -1,1 +1,0 @@
-DOMAIN = "immich_api"
