+++ strings.json
@@ -3,10 +3,9 @@
     "step": {
       "user": {
         "title": "Configure Immich",
-        "description": "Please enter the host and API key to connect to Immich.",
+        "description": "Please enter the host.",
         "data": {
-          "host": "[%key:common::config_flow::data::host%]",
-          "api_key": "[%key:common::config_flow::data::api_key%]"
+          "host": "[%key:common::config_flow::data::host%]"
         }
       }
     },
